requirements.txt
amqp==2.2.2
billiard==3.5.0.3
celery==4.0.2
certifi==2017.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
dj-database-url==0.4.2
Django==1.11.1
django-allauth==0.34.0
django-bootstrap-themes==3.3.6
django-cors-middleware==1.3.1
django-extensions==1.9.7
django-material==1.1.1
django-oauth-toolkit==1.0.0
django-rest-auth==0.9.2
django-rest-swagger==2.0.7
djangorestframework==3.5.1
GeoIP==1.3.2
graphviz==0.8.1
gunicorn==19.7.1
idna==2.6
itypes==1.1.0
Jinja2==2.10
kombu==4.1.0
MarkupSafe==1.0
oauthlib==2.0.6
olefile==0.44
openapi-codec==1.3.2
Pillow==4.3.0
psycopg2==2.7.1
pygraphviz==1.3.1
python3-openid==3.1.0
pytz==2017.3
redis==2.10.5
requests==2.18.4
requests-oauthlib==0.8.0
simplejson==3.13.2
six==1.9.0
typing==3.6.2
uritemplate==3.0.0
urllib3==1.22
vine==1.1.4
whitenoise==3.3.0
